REFACTOR: Create dedicated component for SVG download button.
