+++ packages/frontend/src/visualization/export_svg.ts
