+++ packages/frontend/src/visualization/index.ts
@@ -6,7 +6,8 @@ These tools are applied visualize models in the `analysis` module.
  */
 
 export * from "./types";
-export * from "./export";
+export * from "./export_svg";
+export * from "./export_svg_components";
 export * from "./graph_svg";
 export * from "./graphviz";
 export * from "./graphviz_svg";
